Conflicts:
	ground/openpilotgcs/src/libs/opmapcontrol/src/mapwidget/uavitem.cpp

+++ ground/openpilotgcs/src/plugins/config/cfg_vehicletypes/configfixedwingwidget.cpp
@@ -299,7 +299,7 @@ bool ConfigFixedWingWidget::setupFrameFixedWing(QString airframeType)
 
     int channel;
     //disable all
-    for (channel=0; channel<VehicleConfig::CHANNEL_NUMELEM; channel++)
+    for (channel=0; (unsigned int) channel<VehicleConfig::CHANNEL_NUMELEM; channel++)
     {
         setMixerType(mixer,channel,VehicleConfig::MIXERTYPE_DISABLED);
         resetMixerVector(mixer, channel);
@@ -386,7 +386,7 @@ bool ConfigFixedWingWidget::setupFrameElevon(QString airframeType)
     int channel;
     double value;
     //disable all
-    for (channel=0; channel<VehicleConfig::CHANNEL_NUMELEM; channel++)
+    for (channel=0; (unsigned int) channel<VehicleConfig::CHANNEL_NUMELEM; channel++)
     {
         setMixerType(mixer,channel,VehicleConfig::MIXERTYPE_DISABLED);
         resetMixerVector(mixer, channel);
@@ -471,7 +471,7 @@ bool ConfigFixedWingWidget::setupFrameVtail(QString airframeType)
     int channel;
     double value;
     //disable all
-    for (channel=0; channel<VehicleConfig::CHANNEL_NUMELEM; channel++)
+    for (channel=0; (unsigned int) channel<VehicleConfig::CHANNEL_NUMELEM; channel++)
     {
         setMixerType(mixer,channel,VehicleConfig::MIXERTYPE_DISABLED);
         resetMixerVector(mixer, channel);
